pyproject.toml
[project]
name = "five-day-line"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "akshare>=1.18.22",
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "pandas-stubs>=3.0.0.260204",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]
